Fix roi cooridnates type

+++ tests/test_tracker_executor2.py
@@ -292,7 +292,7 @@ class TrackerExecutorController:
 
 
 NUM_ROIS = 10
-ROI_COORDINATES = cupy.array([[0, 0]] * NUM_ROIS, dtype=cupy.uint16)
+ROI_COORDINATES = cupy.array([[0, 0]] * NUM_ROIS, dtype=cupy.uint32)
 ROI_SIZE = 100
 
 QI_TRACKER_ARGUMENTS = {
